requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
numpy==1.26.3
gunicorn==21.2.0
flask-cors==4.0.0
python-magic==0.4.27
validators==0.22.0
matplotlib
pandas
seaborn
textblob